cellophane: add `#:composite?` argument that is on by default
By default, `cellophane` now applies an alpha to the pict as a whole,
insteda of to the individual drawing commands that compose the pict.
The implementation uses `start-alpha` and `end-alpha` in `dc<%>`.

This is a backward-incompatible change, because it modifies the
rendering of `cellophane` picts. To the degree the this change affects
how picts render, through, it will practicaly always make them look
better. The main drawback is that it can be noticeably slow for
animations involving large images.