add `#:nth` argument to `lt-find`, etc.
Instead of always returning the location for the first found instance
of a pict, allow `#:nth` to select the instance, and report an error
if `#:nth` is `'unique` and difference instances would have different
results.

Also, fix error reporting and the function name to be the actual name,
instead of a `find-XX` placeholder.